dots(emacs): added conf to allow configuration of lsp start dir
eglot uses the root of the projects as denoted by project.el to
specify the start dir of a LSP server. Since we don't use any of
project.el's functionality we're going to use a `.project` file to
denote and subprojects in the box.

This should fix this issue with rust-analyzer failing to start when
working within a monorepo, I've tested with the aoc/2024 project and
it seems to be working fine.